Create README.md
Javascript Create Calendar

The small project to pratice how to set up and make the calendar with JS. 
It help us to check and see calendar in the real time. 
I am begin, if someone have some ideas for my project. 
You can fix and communicate together to make the project that is prefect in the future

You can use my source and it's open sources for everyone